Put together the first stable version of the logic
Even though it sounded simple at first sight, the display
logic caused quite a trouble to get working. It updates itself as
the user presses digit buttons and, whenever an operator is activated,
some auxiliary functions come in handy to cope with the many sorts of
calculations to be made.

A "clear" feature is present as to help out any mistakes made by the
user.

Flexbox played a major role in the overall stylization, hence
the dimensions of the calculator are not affected by external window
changes. Another CSS attributes were tinkered with so that the visuals
would look more appealing.

The most challenging part was to implement a way to let the own
operators behave as an equals sign during operators chaining.